[SERVICE-478] File renaming
