+++ test/demo/raiseIntent.test.ts
@@ -1,7 +1,7 @@
 import 'jest';
 
 import {ResolveError} from '../../src/client/errors';
-import {INTENT_LISTENER_TIMEOUT} from '../../src/provider/model/implementation/Fin';
+import {INTENT_LISTENER_TIMEOUT} from '../../src/provider/model/FinEnvironment';
 
 import {fin} from './utils/fin';
 import * as fdc3Remote from './utils/fdc3RemoteExecution';
